// src/app/models/situation-interface.ts
export interface SituationInterface {
  _id: string;
  name: string;
  user: any;
  title: string
  titleEn: string
  description: string;
  descriptionEn: string;
  order: number;
  date: string;
  enable: boolean;
  status: number;
  createdAt: string;
  callNote: boolean
  updatedAt: string;
  __v: number;
}